refactor(__init__.py): relocate functions in root __init__.py
This relocates the functions to ther places in the code, so that the root __init__.py is primarily
used for determining which functions should be exported.
