+++ trame/dev/__init__.py
@@ -1,0 +1,5 @@
+from .dev import main
+
+__all__ = [
+    "main",
+]
