+++ trame/layouts/__init__.py
@@ -1,6 +1,15 @@
-from trame.layouts.core import (
+from .core import (
     AbstractLayout,
     FullScreenPage,
     SinglePage,
     SinglePageWithDrawer,
+    update_layout,
 )
+
+__all__ = [
+    "AbstractLayout",
+    "FullScreenPage",
+    "SinglePage",
+    "SinglePageWithDrawer",
+    "update_layout",
+]
