+++ trame/state/__init__.py
@@ -1,6 +1,11 @@
-from .state import get_state, update_state
+from .decorators import change
+from .state import flush_state, get_state, is_dirty, is_dirty_all, update_state
 
 __all__ = [
+    "change",
+    "flush_state",
     "get_state",
+    "is_dirty",
+    "is_dirty_all",
     "update_state",
 ]
