+++ trame/trigger/__init__.py
@@ -1,0 +1,7 @@
+from .decorators import trigger
+from .trigger import trigger_key
+
+__all__ = [
+    "trigger",
+    "trigger_key",
+]
