+++ trame/utils/__init__.py
@@ -1,14 +1,19 @@
-from .app_server_thread import AppServerThread
-from .client_window_process import ClientWindowProcess
+from .desktop import AppServerThread, ClientWindowProcess
+from .cli import get_cli_parser
 from .compose import compose_callbacks
-from .logging import log_js_error
+from .filesystem import base_directory
+from .logging import log_js_error, print_server_info, validate_key_names
 from .version import get_version
 
 
 __all__ = [
-    'AppServerThread',
-    'ClientWindowProcess',
-    'compose_callbacks',
-    'get_version',
-    'log_js_error',
+    "AppServerThread",
+    "base_directory",
+    "ClientWindowProcess",
+    "compose_callbacks",
+    "get_cli_parser",
+    "get_version",
+    "log_js_error",
+    "print_server_info",
+    "validate_key_names",
 ]
